worker/pyproject.toml
[project]
name = "worker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "essentia>=2.1b6.dev1389",
    "ffmpeg-python>=0.2.0",
]
